requirements.txt
flask
flask-cors
PyPDF2
sentence-transformers
faiss-cpu
numpy
google-generativeai
gunicorn
